split and reorganize emote input
